fix(conformance): restore aggregate gate
