+++ crates/tsz-core/src/config/lib_resolution.rs
@@ -710,12 +710,16 @@ pub fn is_known_lib_name(lib_name: &str) -> bool {
     if normalized.is_empty() {
         return false;
     }
-    if let Ok(lib_dir) = default_lib_dir()
-        && let Ok(map) = build_lib_map(&lib_dir)
+    let disk_lib_map = default_lib_dir()
+        .ok()
+        .and_then(|lib_dir| build_lib_map(&lib_dir).ok());
+    if disk_lib_map
+        .as_ref()
+        .is_some_and(|map| map.contains_key(&normalized))
     {
-        if map.contains_key(&normalized) {
-            return true;
-        }
+        return true;
+    }
+    if disk_lib_map.is_some() {
         // The on-disk TypeScript installation may predate this lib name
         // (e.g. `esnext.disposable` and `esnext.float16` were added in TS 5.8).
         // Fall through to the embedded catalog so tsz's own snapshot is always
